Revert "Rename test_pandas_datasource.py to test_pandas_filesystem_datasource.py"
This reverts commit d0da1a2b1216d95f3ca3a02a080a25593ea18c88.
